# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LandMate API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export type AnalyzeDocumentRequestDocumentType =
  (typeof AnalyzeDocumentRequestDocumentType)[keyof typeof AnalyzeDocumentRequestDocumentType];

export const AnalyzeDocumentRequestDocumentType = {
  indenture: "indenture",
  titleDeed: "titleDeed",
  lease: "lease",
  glcForm: "glcForm",
  unknown: "unknown",
} as const;

export type AnalyzeDocumentRequestMode =
  (typeof AnalyzeDocumentRequestMode)[keyof typeof AnalyzeDocumentRequestMode];

export const AnalyzeDocumentRequestMode = {
  explain: "explain",
  redFlags: "redFlags",
} as const;

export type AnalyzeDocumentRequestLanguage =
  (typeof AnalyzeDocumentRequestLanguage)[keyof typeof AnalyzeDocumentRequestLanguage];

export const AnalyzeDocumentRequestLanguage = {
  english: "english",
  twi: "twi",
  pidgin: "pidgin",
} as const;

export interface AnalyzeDocumentRequest {
  /** Extracted text from the land document */
  documentText: string;
  documentType: AnalyzeDocumentRequestDocumentType;
  mode: AnalyzeDocumentRequestMode;
  language?: AnalyzeDocumentRequestLanguage;
  /** Original file name for display */
  fileName?: string;
}

export type AnalyzeDocumentResponseOverallRiskLevel =
  (typeof AnalyzeDocumentResponseOverallRiskLevel)[keyof typeof AnalyzeDocumentResponseOverallRiskLevel];

export const AnalyzeDocumentResponseOverallRiskLevel = {
  LOW: "LOW",
  MEDIUM: "MEDIUM",
  HIGH: "HIGH",
} as const;

export interface ImportantDate {
  label: string;
  value: string;
  daysUntil?: number | null;
}

export type RedFlagSeverity =
  (typeof RedFlagSeverity)[keyof typeof RedFlagSeverity];

export const RedFlagSeverity = {
  HIGH: "HIGH",
  MEDIUM: "MEDIUM",
  LOW: "LOW",
} as const;

export interface RedFlag {
  severity: RedFlagSeverity;
  issue: string;
  explanation: string;
  recommendation: string;
}

export interface AnalyzeDocumentResponse {
  summary: string;
  keyPoints: string[];
  importantDates?: ImportantDate[];
  nextSteps: string[];
  redFlags?: RedFlag[];
  overallRiskLevel?: AnalyzeDocumentResponseOverallRiskLevel;
  documentType?: string;
  disclaimer: string;
}

export type ProcessGuidanceRequestLanguage =
  (typeof ProcessGuidanceRequestLanguage)[keyof typeof ProcessGuidanceRequestLanguage];

export const ProcessGuidanceRequestLanguage = {
  english: "english",
  twi: "twi",
  pidgin: "pidgin",
} as const;

export interface ProcessGuidanceRequest {
  processName: string;
  language?: ProcessGuidanceRequestLanguage;
}

export interface ProcessStep {
  stepNumber: number;
  title: string;
  description: string;
  office?: string | null;
  fee?: string | null;
  documents?: string[];
}

export interface ProcessGuidanceResponse {
  processName: string;
  steps: ProcessStep[];
  requiredDocuments?: string[];
  estimatedFees?: string;
  estimatedDuration?: string;
  offices?: string[];
  disclaimer: string;
}

export type FormAssistanceRequestLanguage =
  (typeof FormAssistanceRequestLanguage)[keyof typeof FormAssistanceRequestLanguage];

export const FormAssistanceRequestLanguage = {
  english: "english",
  twi: "twi",
  pidgin: "pidgin",
} as const;

export interface FormAssistanceRequest {
  documentText: string;
  documentType: string;
  language?: FormAssistanceRequestLanguage;
}

export interface FieldGuidance {
  fieldName: string;
  instructions: string;
  example?: string | null;
  required: boolean;
}

export interface FormAssistanceResponse {
  formName: string;
  fieldGuidance: FieldGuidance[];
  commonMistakes?: string[];
  summary: string;
  disclaimer: string;
}

export type ChatMessageRole =
  (typeof ChatMessageRole)[keyof typeof ChatMessageRole];

export const ChatMessageRole = {
  user: "user",
  assistant: "assistant",
} as const;

export interface ChatMessage {
  role: ChatMessageRole;
  content: string;
}

export type ChatRequestLanguage =
  (typeof ChatRequestLanguage)[keyof typeof ChatRequestLanguage];

export const ChatRequestLanguage = {
  english: "english",
  twi: "twi",
  pidgin: "pidgin",
} as const;

export interface ChatRequest {
  message: string;
  conversationHistory: ChatMessage[];
  documentContext?: string | null;
  documentType?: string | null;
  language?: ChatRequestLanguage;
}

export interface ChatResponse {
  response: string;
  disclaimer: string;
}