# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LandMate API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Analyze a document using Claude AI for explanation or red flag detection
 * @summary Analyze a land document
 */
export const analyzeDocumentBodyLanguageDefault = `english`;

export const AnalyzeDocumentBody = zod.object({
  documentText: zod.string().describe("Extracted text from the land document"),
  documentType: zod.enum([
    "indenture",
    "titleDeed",
    "lease",
    "glcForm",
    "unknown",
  ]),
  mode: zod.enum(["explain", "redFlags"]),
  language: zod
    .enum(["english", "twi", "pidgin"])
    .default(analyzeDocumentBodyLanguageDefault),
  fileName: zod.string().optional().describe("Original file name for display"),
});

export const AnalyzeDocumentResponse = zod.object({
  summary: zod.string(),
  keyPoints: zod.array(zod.string()),
  importantDates: zod
    .array(
      zod.object({
        label: zod.string(),
        value: zod.string(),
        daysUntil: zod.number().nullish(),
      }),
    )
    .optional(),
  nextSteps: zod.array(zod.string()),
  redFlags: zod
    .array(
      zod.object({
        severity: zod.enum(["HIGH", "MEDIUM", "LOW"]),
        issue: zod.string(),
        explanation: zod.string(),
        recommendation: zod.string(),
      }),
    )
    .optional(),
  overallRiskLevel: zod.enum(["LOW", "MEDIUM", "HIGH"]).optional(),
  documentType: zod.string().optional(),
  disclaimer: zod.string(),
});

/**
 * Get step-by-step guidance for a GLC process
 * @summary Get GLC process guidance
 */
export const getProcessGuidanceBodyLanguageDefault = `english`;

export const GetProcessGuidanceBody = zod.object({
  processName: zod.string(),
  language: zod
    .enum(["english", "twi", "pidgin"])
    .default(getProcessGuidanceBodyLanguageDefault),
});

export const GetProcessGuidanceResponse = zod.object({
  processName: zod.string(),
  steps: zod.array(
    zod.object({
      stepNumber: zod.number(),
      title: zod.string(),
      description: zod.string(),
      office: zod.string().nullish(),
      fee: zod.string().nullish(),
      documents: zod.array(zod.string()).optional(),
    }),
  ),
  requiredDocuments: zod.array(zod.string()).optional(),
  estimatedFees: zod.string().optional(),
  estimatedDuration: zod.string().optional(),
  offices: zod.array(zod.string()).optional(),
  disclaimer: zod.string(),
});

/**
 * Get field-by-field guidance for filling a GLC form
 * @summary Get GLC form filling assistance
 */
export const getFormAssistanceBodyLanguageDefault = `english`;

export const GetFormAssistanceBody = zod.object({
  documentText: zod.string(),
  documentType: zod.string(),
  language: zod
    .enum(["english", "twi", "pidgin"])
    .default(getFormAssistanceBodyLanguageDefault),
});

export const GetFormAssistanceResponse = zod.object({
  formName: zod.string(),
  fieldGuidance: zod.array(
    zod.object({
      fieldName: zod.string(),
      instructions: zod.string(),
      example: zod.string().nullish(),
      required: zod.boolean(),
    }),
  ),
  commonMistakes: zod.array(zod.string()).optional(),
  summary: zod.string(),
  disclaimer: zod.string(),
});

/**
 * Ask follow-up questions in context of the analyzed document
 * @summary Chat with LandMate Q&A
 */
export const chatWithLandmateBodyLanguageDefault = `english`;

export const ChatWithLandmateBody = zod.object({
  message: zod.string(),
  conversationHistory: zod.array(
    zod.object({
      role: zod.enum(["user", "assistant"]),
      content: zod.string(),
    }),
  ),
  documentContext: zod.string().nullish(),
  documentType: zod.string().nullish(),
  language: zod
    .enum(["english", "twi", "pidgin"])
    .default(chatWithLandmateBodyLanguageDefault),
});

export const ChatWithLandmateResponse = zod.object({
  response: zod.string(),
  disclaimer: zod.string(),
});