# lib/api-zod/src/generated/types/analyzeDocumentResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LandMate API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AnalyzeDocumentResponseOverallRiskLevel } from "./analyzeDocumentResponseOverallRiskLevel";
import type { ImportantDate } from "./importantDate";
import type { RedFlag } from "./redFlag";

export interface AnalyzeDocumentResponse {
  summary: string;
  keyPoints: string[];
  importantDates?: ImportantDate[];
  nextSteps: string[];
  redFlags?: RedFlag[];
  overallRiskLevel?: AnalyzeDocumentResponseOverallRiskLevel;
  documentType?: string;
  disclaimer: string;
}