# lib/api-zod/src/generated/types/chatRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LandMate API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ChatMessage } from "./chatMessage";
import type { ChatRequestLanguage } from "./chatRequestLanguage";

export interface ChatRequest {
  message: string;
  conversationHistory: ChatMessage[];
  documentContext?: string | null;
  documentType?: string | null;
  language?: ChatRequestLanguage;
}