# lib/api-zod/src/generated/types/chatResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LandMate API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ChatResponse {
  response: string;
  disclaimer: string;
}