# lib/api-zod/src/generated/types/formAssistanceResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LandMate API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FieldGuidance } from "./fieldGuidance";

export interface FormAssistanceResponse {
  formName: string;
  fieldGuidance: FieldGuidance[];
  commonMistakes?: string[];
  summary: string;
  disclaimer: string;
}