# lib/api-zod/src/generated/types/importantDate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LandMate API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ImportantDate {
  label: string;
  value: string;
  daysUntil?: number | null;
}