# lib/api-zod/src/generated/types/processGuidanceRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LandMate API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProcessGuidanceRequestLanguage } from "./processGuidanceRequestLanguage";

export interface ProcessGuidanceRequest {
  processName: string;
  language?: ProcessGuidanceRequestLanguage;
}