# lib/api-zod/src/generated/types/processGuidanceResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LandMate API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProcessStep } from "./processStep";

export interface ProcessGuidanceResponse {
  processName: string;
  steps: ProcessStep[];
  requiredDocuments?: string[];
  estimatedFees?: string;
  estimatedDuration?: string;
  offices?: string[];
  disclaimer: string;
}