# lib/api-zod/src/generated/types/redFlag.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LandMate API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RedFlagSeverity } from "./redFlagSeverity";

export interface RedFlag {
  severity: RedFlagSeverity;
  issue: string;
  explanation: string;
  recommendation: string;
}